Changed RandomGeometry functions to accept ints rather than longs.

+++ DotSpatial.Analysis/RandomGeometry.cs
@@ -29,14 +29,14 @@ namespace DotSpatial.Analysis
         /// <param name="ConstrainingFeatures">Random points will be generated inside all features in this feature set.</param>
         /// <param name="NumberOfPoints">The number of points to be randomly generated.</param>
         /// <returns>A point feature set with the randomly created features.</returns>
-        public static FeatureSet RandomPoints(IFeatureSet ConstrainingFeatures, long NumberOfPoints, ICancelProgressHandler cancelProgressHandler = null)
+        public static FeatureSet RandomPoints(IFeatureSet ConstrainingFeatures, int NumberOfPoints, ICancelProgressHandler cancelProgressHandler = null)
         {
             //This function generates random points within the boundaries of all polygon features in a feature set
             FeatureSet fsOut = new FeatureSet();
             fsOut.FeatureType = FeatureType.Point;
             Coordinate c = new Coordinate();
             Random r = new Random();
-            long i = 0;
+            int i = 0;
             while (i < NumberOfPoints)
             {
                 c = new Coordinate();
@@ -70,14 +70,14 @@ namespace DotSpatial.Analysis
         /// <param name="ConstrainingFeature">Random points will be generated inside this polygon feature.</param>
         /// <param name="NumberOfPoints">The number of points to be randomly generated.</param>
         /// <returns>A point feature set with the randomly created features.</returns>
-        public static FeatureSet RandomPoints(Feature ConstrainingFeature, long NumberOfPoints)
+        public static FeatureSet RandomPoints(Feature ConstrainingFeature, int NumberOfPoints)
         {
             //This function generates random points within the boundaries of one polygon feature
             FeatureSet fsOut = new FeatureSet();
             fsOut.FeatureType = FeatureType.Point;
             Coordinate c = new Coordinate();
             Random r = new Random();
-            long i = 0;
+            int i = 0;
             while (i < NumberOfPoints)
             {
                 c = new Coordinate();
